requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
sentence-transformers==2.2.2
faiss-cpu==1.13.2
anthropic==0.100.0
python-dotenv==1.0.0
lxml==4.9.3
numpy==1.26.4
